+++ backend/credit-service/Models/CreditRate.cs
@@ -6,7 +6,7 @@ namespace credit_service.Models
 		public Guid CreditRateId { get; set; }
 		public string Title { get; set; }
 		public string? Description { get; set; }
-		public int InterestRate { get; set; }
+		public double InterestRate { get; set; }
 		public CreditRateStatus IsActive { get; set; }
 
         public CreditRate()
